jpa models updated to match deploy db. fixed bugs created when updating the models

+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/SongDao.java
@@ -11,7 +11,7 @@ public interface SongDao extends CrudDao<Song> {
     List<Song> findByTitleContaining(String sub);
 
     int saveSongArtist(Song song, Artist artist);
-    boolean updateRating(long songId, float newRating, int newRatingAmount);
+    boolean updateRating(long songId, double newRating, int newRatingAmount);
     boolean hasUserReviewed(long userId, long songId);
 
     boolean deleteReviewsFromSong(long songId);
